numerical characteristics completed
NAS Number of assignment statements in the algorithm.
LoC Lines of code.
N1 Total number of operators in the algorithm.
N2 Total number of operands in the algorithm.
n1 Number of unique operators in the algorithm.
n2 Number of unique operands in the algorithm.
N Program length (N = N1 + N2).
n Program vocabulary (n = n1 + n2).
NoV Number of variables in the algorithm.
NoL Number of loops. Supported loops are for loop, while
loop and do while loop.
NoNL Number of nested loops in the algorithm.
NoB Number of blocks in the algorithm. A block refers to a
sequence of statements wrapped in curly braces, for example,
a method or a control structure (loops and conditionals).